+++ src/treinador/RepositorioTreinadores.java
@@ -1,9 +1,9 @@
 package treinador;
 
 public interface RepositorioTreinadores  {
-    public void inserir (String nome, int idade, double salario, String registroFifa, int experiencia);
+    public void inserir (String nome, int idade, double salario, String registroFifa, double experiencia);
 
-    public void atualizar (String registroFifa, int experiencia, int idade, double salario);
+    public void atualizar (String registroFifa, double experiencia, int idade, double salario);
 
     public void remover (String registroFifa);
